requirements.txt
websockets
asyncio
boto3
requests
python-dotenv
pyinquirer